+++ test/WebJobs.Script.Tests.Integration/ContainerManagement/LinuxContainerInitializationHostedServiceTests.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Script.WebHost;
@@ -24,16 +23,14 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Integration.ContainerManagement
 {
     [Trait(TestTraits.Category, TestTraits.EndToEnd)]
     [Trait(TestTraits.Group, TestTraits.ContainerInstanceTests)]
-    public class LegionContainerInitializationHostServiceTests : IDisposable
+    public class LinuxContainerInitializationHostedServiceTests : IDisposable
     {
         private const string ContainerStartContextUri = "https://containerstartcontexturi";
-        private const string ContextFile = "Context.txt";
-        private const string ContainerSpecializationContextPath = "/container-specialization-context";
         private readonly Mock<IInstanceManager> _instanceManagerMock;
         private readonly StartupContextProvider _startupContextProvider;
         private readonly TestEnvironment _environment;
 
-        public LegionContainerInitializationHostServiceTests()
+        public LinuxContainerInitializationHostedServiceTests()
         {
             _instanceManagerMock = new Mock<IInstanceManager>(MockBehavior.Strict);
 
@@ -52,20 +49,20 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Integration.ContainerManagement
             _environment.SetEnvironmentVariable(AzureWebsiteInstanceId, null);
             Assert.False(_environment.IsAnyLinuxConsumption());
 
-            var initializationHostService = new LegionContainerInitializationHostService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostService>.Instance, _startupContextProvider);
+            var initializationHostService = new AtlasContainerInitializationHostedService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostedService>.Instance, _startupContextProvider);
             await initializationHostService.StartAsync(CancellationToken.None);
         }
 
         [Fact]
-        public async Task Does_Not_Run_In_Linux_Container_On_Atlas()
+        public async Task Does_Not_Run_In_Linux_Container_On_Legion()
         {
             _environment.SetEnvironmentVariable(ContainerName, "abcd");
-            _environment.SetEnvironmentVariable(LegionServiceHost, null);
+            _environment.SetEnvironmentVariable(LegionServiceHost, "1");
             Assert.True(_environment.IsAnyLinuxConsumption());
-            Assert.True(_environment.IsLinuxConsumptionOnAtlas());
-            Assert.False(_environment.IsLinuxConsumptionOnLegion());
+            Assert.False(_environment.IsLinuxConsumptionOnAtlas());
+            Assert.True(_environment.IsFlexConsumptionSku());
 
-            var initializationHostService = new LegionContainerInitializationHostService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostService>.Instance, _startupContextProvider);
+            var initializationHostService = new AtlasContainerInitializationHostedService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostedService>.Instance, _startupContextProvider);
             await initializationHostService.StartAsync(CancellationToken.None);
         }
 
@@ -93,10 +90,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Integration.ContainerManagement
             var encryptedHostAssignmentContext = GetEncryptedHostAssignmentContext(hostAssignmentContext, containerEncryptionKey);
             var serializedContext = JsonConvert.SerializeObject(new { encryptedContext = encryptedHostAssignmentContext });
 
-            // set start context in the file
-            CreateTemporaryStartContextFile(serializedContext);
-
-            _environment.SetEnvironmentVariable(ContainerSpecializationContextVolumePath, ContainerSpecializationContextPath);
+            _environment.SetEnvironmentVariable(ContainerStartContext, serializedContext);
             _environment.SetEnvironmentVariable(ContainerEncryptionKey, containerEncryptionKey);
             AddLinuxConsumptionSettings(_environment);
 
@@ -106,28 +100,75 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Integration.ContainerManagement
 
             _instanceManagerMock.Setup(manager => manager.StartAssignment(It.Is<HostAssignmentContext>(context => hostAssignmentContext.Equals(context) && !context.IsWarmupRequest))).Returns(true);
 
-            var initializationHostService = new LegionContainerInitializationHostService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostService>.Instance, _startupContextProvider);
+            var initializationHostService = new AtlasContainerInitializationHostedService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostedService>.Instance, _startupContextProvider);
             await initializationHostService.StartAsync(CancellationToken.None);
 
-            _instanceManagerMock.Verify(m => m.SpecializeMSISidecar(It.Is<HostAssignmentContext>(context => hostAssignmentContext.Equals(context) && !context.IsWarmupRequest)), Times.Never);
-            
+            _instanceManagerMock.Verify(m =>
+                m.SpecializeMSISidecar(It.Is<HostAssignmentContext>(context =>
+                    hostAssignmentContext.Equals(context) && !context.IsWarmupRequest)), Times.Once);
+
             _instanceManagerMock.Verify(manager => manager.StartAssignment(It.Is<HostAssignmentContext>(context => hostAssignmentContext.Equals(context) && !context.IsWarmupRequest)), Times.Once);
 
             var hostSecrets = _startupContextProvider.GetHostSecretsOrNull();
             Assert.Equal("test-key", hostSecrets.MasterKey);
-            DeleteTemporaryStartContextFile();
         }
-
 
         [Fact]
         public async Task Does_Not_Assign_If_Context_Not_Available()
         {
-            var initializationHostService = new LegionContainerInitializationHostService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostService>.Instance, _startupContextProvider);
+            var initializationHostService = new AtlasContainerInitializationHostedService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostedService>.Instance, _startupContextProvider);
             await initializationHostService.StartAsync(CancellationToken.None);
 
+            _instanceManagerMock.Verify(m => m.SpecializeMSISidecar(It.IsAny<HostAssignmentContext>()), Times.Never);
             _instanceManagerMock.Verify(manager => manager.StartAssignment(It.IsAny<HostAssignmentContext>()), Times.Never);
         }
 
+        [Fact]
+        public async Task Assigns_Even_If_MSI_Specialization_Fails()
+        {
+            var containerEncryptionKey = TestHelpers.GenerateKeyHexString();
+            var hostAssignmentContext = GetHostAssignmentContext();
+            var secrets = new FunctionAppSecrets();
+            secrets.Host = new FunctionAppSecrets.HostSecrets
+            {
+                Master = "test-key",
+                Function = new Dictionary<string, string>
+                {
+                    { "host-function-key-1", "test-key" }
+                },
+                System = new Dictionary<string, string>
+                {
+                    { "host-system-key-1", "test-key" }
+                }
+            };
+            hostAssignmentContext.Secrets = secrets;
+            hostAssignmentContext.MSIContext = new MSIContext();
+
+            var encryptedHostAssignmentContext = GetEncryptedHostAssignmentContext(hostAssignmentContext, containerEncryptionKey);
+            var serializedContext = JsonConvert.SerializeObject(new { encryptedContext = encryptedHostAssignmentContext });
+
+            _environment.SetEnvironmentVariable(ContainerStartContext, serializedContext);
+            _environment.SetEnvironmentVariable(ContainerEncryptionKey, containerEncryptionKey);
+            AddLinuxConsumptionSettings(_environment);
+
+            _instanceManagerMock.Setup(m =>
+                m.SpecializeMSISidecar(It.Is<HostAssignmentContext>(context =>
+                    hostAssignmentContext.Equals(context) && !context.IsWarmupRequest))).Returns(Task.FromResult("msi-specialization-error"));
+
+            _instanceManagerMock.Setup(manager => manager.StartAssignment(It.Is<HostAssignmentContext>(context => hostAssignmentContext.Equals(context) && !context.IsWarmupRequest))).Returns(true);
+
+            var initializationHostService = new AtlasContainerInitializationHostedService(_environment, _instanceManagerMock.Object, NullLogger<LinuxContainerInitializationHostedService>.Instance, _startupContextProvider);
+            await initializationHostService.StartAsync(CancellationToken.None);
+
+            _instanceManagerMock.Verify(m =>
+                m.SpecializeMSISidecar(It.Is<HostAssignmentContext>(context =>
+                    hostAssignmentContext.Equals(context) && !context.IsWarmupRequest)), Times.Once);
+
+            _instanceManagerMock.Verify(manager => manager.StartAssignment(It.Is<HostAssignmentContext>(context => hostAssignmentContext.Equals(context) && !context.IsWarmupRequest)), Times.Once);
+
+            var hostSecrets = _startupContextProvider.GetHostSecretsOrNull();
+            Assert.Equal("test-key", hostSecrets.MasterKey);
+        }
 
         private static string GetEncryptedHostAssignmentContext(HostAssignmentContext hostAssignmentContext, string containerEncryptionKey)
         {
@@ -153,27 +194,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Integration.ContainerManagement
         {
             environment.SetEnvironmentVariable(AzureWebsiteInstanceId, string.Empty);
             environment.SetEnvironmentVariable(ContainerName, "ContainerName");
-            environment.SetEnvironmentVariable(LegionServiceHost, "1");
-        }
-        
-        private static void CreateTemporaryStartContextFile(string content)
-        {
-            var contextFilePath = Path.Combine(ContainerSpecializationContextPath, ContextFile);
-            Directory.CreateDirectory(ContainerSpecializationContextPath);
-            File.WriteAllText(contextFilePath, content);
-        }
-
-        private static void DeleteTemporaryStartContextFile()
-        {
-            var contextFilePath = Path.Combine(ContainerSpecializationContextPath, ContextFile);
-            if (File.Exists(contextFilePath))
-            {
-                File.Delete(contextFilePath);
-            }
-            if (Directory.Exists(ContainerSpecializationContextPath))
-            {
-                Directory.Delete(ContainerSpecializationContextPath);
-            }
         }
 
         public void Dispose()
